# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Crypto Trading Bot API for Bybit
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /positions:
    get:
      operationId: listPositions
      tags: [positions]
      summary: List open positions
      parameters:
        - in: query
          name: mode
          schema:
            type: string
            enum: [paper, live, all]
          required: false
      responses:
        "200":
          description: Open positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
    post:
      operationId: openPosition
      tags: [positions]
      summary: Open a manual position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionInput"
      responses:
        "201":
          description: Position opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trade history
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          required: false
        - in: query
          name: mode
          schema:
            type: string
            enum: [paper, live, all]
          required: false
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
...
  /market/ticker:
    get:
      operationId: getMarketTicker
      tags: [market]
      summary: Get market ticker data
      parameters:
        - in: query
          name: symbol
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Ticker data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticker"
  /market/klines:
    get:
      operationId: getKlines
      tags: [market]
      summary: Get OHLCV candlestick data
      parameters:
        - in: query
          name: symbol
          schema:
            type: string
          required: true
        - in: query
          name: interval
          schema:
            type: string
            enum: ["1", "5", "15", "60", "240", "D"]
          required: false
        - in: query
          name: limit
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Candlestick data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kline"
  /market/levels:
    get:
      operationId: getStrategyLevels
      tags: [market]
      summary: Get strategy-calculated price levels for a symbol
      parameters:
        - in: query
          name: symbol
          schema:
            type: string
          required: true
        - in: query
          name: interval
          schema:
            type: string
          required: false
        - in: query
          name: strategies
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Strategy levels for the symbol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyLevels"
  /signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List recent trading signals
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          required: false
      responses:
        "200":
          description: Recent signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
...
  /strategy-presets:
    get:
      operationId: listStrategyPresets
      tags: [bot]
      summary: List all strategy presets
      responses:
        "200":
          description: List of strategy presets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StrategyPreset"
    post:
      operationId: createStrategyPreset
      tags: [bot]
      summary: Create a new strategy preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyPresetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyPreset"
...
components:
  schemas:
...
    Position:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        currentPrice:
          type: number
        quantity:
          type: number
        leverage:
          type: integer
        unrealizedPnl:
          type: number
        unrealizedPnlPercent:
          type: number
        strategy:
          type: string
        mode:
          type: string
          enum: [paper, live]
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        openedAt:
          type: string
        bybitOrderId:
          type: ["string", "null"]
        presetName:
          type: ["string", "null"]
      required:
        - id
        - symbol
        - side
        - entryPrice
        - currentPrice
        - quantity
        - leverage
        - unrealizedPnl
        - unrealizedPnlPercent
        - strategy
        - mode
        - openedAt
    PositionInput:
      type: object
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        quantity:
          type: number
        leverage:
          type: integer
        mode:
          type: string
          enum: [paper, live]
        stopLoss:
          type: number
        takeProfit:
          type: number
      required:
        - symbol
        - side
        - quantity
        - leverage
        - mode
    Trade:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: number
        quantity:
          type: number
        leverage:
          type: integer
        pnl:
          type: number
        pnlPercent:
          type: number
        strategy:
          type: string
        mode:
          type: string
          enum: [paper, live]
        openedAt:
          type: string
        closedAt:
          type: string
        bybitOrderId:
          type: ["string", "null"]
        wasAveraged:
          type: boolean
        averageCount:
          type: integer
        presetName:
          type: ["string", "null"]
      required:
        - id
        - symbol
        - side
        - entryPrice
        - exitPrice
        - quantity
        - leverage
        - pnl
        - pnlPercent
        - strategy
        - mode
        - wasAveraged
        - averageCount
        - openedAt
        - closedAt
...
    Ticker:
      type: object
      properties:
        symbol:
          type: string
        lastPrice:
          type: number
        change24h:
          type: number
        changePercent24h:
          type: number
        volume24h:
          type: number
        high24h:
          type: number
        low24h:
          type: number
      required:
        - symbol
        - lastPrice
        - change24h
        - changePercent24h
        - volume24h
        - high24h
        - low24h
    Kline:
      type: object
      properties:
        timestamp:
          type: integer
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
      required:
        - timestamp
        - open
        - high
        - low
        - close
        - volume
    StrategyLevels:
      type: object
      properties:
        symbol:
          type: string
        currentPrice:
          type: number
        poc:
          type: ["number", "null"]
        fibHigh:
          type: ["number", "null"]
        fibLow:
          type: ["number", "null"]
        fibLevels:
          type: array
          items:
            type: object
            properties:
              ratio:
                type: number
              price:
                type: number
              label:
                type: string
            required: [ratio, price, label]
        orderBlocks:
          type: array
          items:
            type: object
            properties:
              low:
                type: number
              high:
                type: number
              direction:
                type: string
                enum: [long, short]
            required: [low, high, direction]
        openPositions:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              side:
                type: string
              entryPrice:
                type: number
              quantity:
                type: number
              leverage:
                type: integer
              unrealizedPnl:
                type: number
              stopLoss:
                type: ["number", "null"]
              takeProfit:
                type: ["number", "null"]
            required: [id, side, entryPrice, quantity, leverage, unrealizedPnl]
        recentSignals:
          type: array
          items:
            type: object
            properties:
              strategy:
                type: string
              direction:
                type: string
              price:
                type: number
              strength:
                type: number
              createdAt:
                type: string
            required: [strategy, direction, price, strength, createdAt]
        rsiValue:
          type: ["number", "null"]
        rsiValues:
          type: array
          items:
            type: object
            properties:
              time:
                type: number
              value:
                type: number
            required: [time, value]
        volumeProfile:
          type: array
          items:
            type: object
            properties:
              price:
                type: number
              volume:
                type: number
              relVol:
                type: number
            required: [price, volume, relVol]
      required:
        - symbol
        - currentPrice
        - fibLevels
        - orderBlocks
        - openPositions
        - recentSignals
    Signal:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        strategy:
          type: string
        direction:
          type: string
          enum: [long, short, neutral]
        strength:
          type: number
        price:
          type: number
        description:
          type: string
        acted:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - symbol
        - strategy
        - direction
        - strength
        - price
        - description
        - acted
        - createdAt
...
    StrategyPreset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        enabled:
          type: boolean
        symbols:
          type: array
          items:
            type: string
        strategies:
          type: array
          items:
            type: string
        strategyMode:
          type: string
          enum: [AND, OR]
        positionSizeUsdt:
          type: number
        leverage:
          type: integer
        maxPositions:
          type: integer
        stopLossUsdt:
          type: number
        takeProfitUsdt:
          type: number
        averagingEnabled:
          type: boolean
        averagingThresholdPercent:
          type: number
        maxAveragingCount:
          type: integer
        averagingAmountUsdt:
          type: number
        timeframe:
          type: string
        volumeProfileParams:
          type: object
        fibonacciParams:
          type: object
        orderBlockParams:
          type: object
        rsiParams:
          type: object
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - enabled
        - symbols
        - strategies
        - strategyMode
        - positionSizeUsdt
        - leverage
        - maxPositions
        - stopLossUsdt
        - takeProfitUsdt
        - averagingEnabled
        - averagingThresholdPercent
        - maxAveragingCount
        - averagingAmountUsdt
        - timeframe
        - createdAt
        - updatedAt
    StrategyPresetInput:
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
        symbols:
          type: array
          items:
            type: string
        strategies:
          type: array
          items:
            type: string
        strategyMode:
          type: string
          enum: [AND, OR]
        positionSizeUsdt:
          type: number
        leverage:
          type: integer
        maxPositions:
          type: integer
        stopLossUsdt:
          type: number
        takeProfitUsdt:
          type: number
        averagingEnabled:
          type: boolean
        averagingThresholdPercent:
          type: number
        maxAveragingCount:
          type: integer
        averagingAmountUsdt:
          type: number
        timeframe:
          type: string
        volumeProfileParams:
          type: object
        fibonacciParams:
          type: object
        orderBlockParams:
          type: object
        rsiParams:
          type: object
      required:
        - name