# lib/api-zod/src/generated/types/positionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Bot API for Bybit
 * OpenAPI spec version: 0.1.0
 */
import type { PositionInputMode } from './positionInputMode';
import type { PositionInputSide } from './positionInputSide';

export interface PositionInput {
  symbol: string;
  side: PositionInputSide;
  quantity: number;
  leverage: number;
  mode: PositionInputMode;
  stopLoss?: number;
  takeProfit?: number;
}